Banking/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==3.7.1
async-timeout==4.0.3
attrs==25.3.0
black==24.4.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
coverage==7.9.2
dataclasses-json==0.6.7
distro==1.9.0
exceptiongroup==1.3.0
faiss-cpu==1.7.4
flake8==6.0.0
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
importlib-metadata==6.11.0
iniconfig==2.1.0
isort==5.12.0
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.68
langchain-openai==0.3.27
langchain-text-splitters==0.3.8
langgraph==0.5.2
langgraph-checkpoint==2.1.0
langgraph-prebuilt==0.5.2
langgraph-sdk==0.1.72
langsmith==0.4.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mccabe==0.7.0
mdurl==0.1.2
multidict==6.6.3
mypy==1.7.0
mypy_extensions==1.1.0
narwhals==1.46.0
numpy==1.26.3
openai==1.86.0
orjson==3.10.18
ormsgpack==1.10.0
packaging==23.2
pandas==2.1.4
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.8
pluggy==1.6.0
propcache==0.3.2
protobuf==4.25.8
psycopg2-binary==2.9.7
pyarrow==20.0.0
pycodestyle==2.10.0
pydantic==2.7.4
pydantic-settings==2.10.1
pydantic_core==2.18.4
pydeck==0.9.1
pyflakes==3.0.1
Pygments==2.19.2
pytest==7.4.0
pytest-cov==4.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.26.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.23
sqlparse==0.4.4
streamlit==1.28.0
tenacity==8.5.0
tiktoken==0.9.0
toml==0.10.2
tomli==2.2.1
tornado==6.5.1
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
validators==0.35.0
watchdog==6.0.0
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
